requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-core
faiss-cpu
python-dotenv
pypdf
python-docx
pandas
openpyxl